🔀 Feature/project support (#203)
* 🚧 Basic projects page

* 🚧 Migrate to GraphQL so we can use images

* ✨ Add stars and fork count

* 🐛 Only query public repos

* 💄 Changed the style of the repo card

* 🐛 Fixed a bug where a repo without any language tags would display null

* 💄 Updated the list style on different pages

* 🚧 Per project page

* 🎨 Remove check from RepositoryCard

* 💄 Fixed small size difference between having an image or not

* 💄 Improve printing of the page

* ✨ Allow ignoring/hiding projects

* ✨ Add owner to repo

* 🧐 Couple of small changes
- More async tasks in osm.json.ts
- Added doc to schema.prisma

* 🚚 svelte-migrate: renamed files

* ⬆️ svelte-migrate: updated files

* ⬆️ Updated package.json for new version of svelte kit

* 🚚 Rename folder to make it more consistant

* 🐛 Fix error for loading Github Activity


* 💄 Improved spacing of the date section

* 🚧 Change database for new page system

* 🚧 Basic project page support

* 🚧 Generate basic page if no custom page is there

* 🐛 Case insensitive filtering

* 🚚 Add owner to URL

* 🚧 Add link to GitHub

* 💄 Added small variant of button

* 💄 Only change BG for large button

* ⬆️ Updated packages

* ✨ Changed the way the project pages are handled

* 💄 Changed particle loading animation

* 💄 Fixed loading animation on project page

* 🚨 Fixed formatter issues

* Update src/routes/[lang]/projects/[owner]/[project]/+page.server.ts

+++ src/routes/api/github.json/+server.ts
@@ -2,7 +2,6 @@ import type { GithubEvent } from '$lib/@types/github';
 import {
 	checkCacheState,
 	extractLangFromUrl,
-	getCacheData,
 	saveCacheData,
 	updateCacheState,
 	translateCache,
@@ -11,9 +10,12 @@ import {
 import { GitHubUsername } from '$lib/common';
 import { loadTranslations } from '$lib/i18n';
 import { prisma } from '$lib/prisma';
-import type { RequestHandler } from './__types/github.json';
+import type { RequestHandler } from './$types';
 
 export const GET: RequestHandler = async ({ url }) => {
+	// Load the required translations
+	await loadTranslations(extractLangFromUrl(url), '/api/github');
+
 	// Get information about the cache
 	const { id, cacheState } = await checkCacheState('github');
 
@@ -26,6 +28,10 @@ export const GET: RequestHandler = async ({ url }) => {
 				`https://api.github.com/users/${GitHubUsername}/events/public?per_page=10`
 			);
 			const json: GithubEvent<any>[] = await response.json(); // eslint-disable-line @typescript-eslint/no-explicit-any
+
+			if (response.status != 200) {
+				throw 'Faulty response';
+			}
 
 			// Remove old cache
 			await prisma.remoteData.deleteMany({
@@ -313,15 +319,11 @@ export const GET: RequestHandler = async ({ url }) => {
 			// Update the cache state
 			await updateCacheState(id);
 		} catch {
-			return {
-				status: 500,
-				...(await getCacheData(id))
-			};
+			return new Response(JSON.stringify(await translateCache(id)), {
+				status: 500
+			});
 		}
 	}
 
-	// Load the required translations
-	await loadTranslations(extractLangFromUrl(url), '/api/github');
-
-	return await translateCache(id);
+	return new Response(JSON.stringify(await translateCache(id)));
 };
